Delete redundant .github/CONTRIBUTING.md
This contributing guide doesn’t talk about the LessWrong 2.0 project, but about the template that this project is based on. There is already a newer CONTRIBUTING.md in the project root.

This version of CONTRIBUTING.md, in .github, is the one that GitHub links new contributors to when they visit the Issues page.