Add margin to text on mobile
# Description
The text of articles was right up to the edge of the screen on mobile which made text reading ever so slightly annoying. This will give the text some room to breath :)

# Screenshots
Before:
![image](https://user-images.githubusercontent.com/4751193/36637368-7ff5a66a-198f-11e8-8159-e4a8b021f44d.png)

After:
!![image](https://user-images.githubusercontent.com/4751193/36637378-a920b6b0-198f-11e8-8710-5ac68d0755e9.png)

# Testing
I couldn't get the dev environment working so this is untested other than fiddling the css on the page.