Send MathJax's stylesheet only once
Before, we would have MathJax embed its stylesheet into every LaTeX
element, meaning that posts with a lot of LaTeX elements would have a
lot of copies of its stylesheet. Since the stylesheet is 19kb and LaTeX
elements can be quite small, this adds up to a really surprising amount
of bloat - 1.5MB of the front page's GraphQL query result sent to the
client, as of when I checked the main site today.

This patch adds MathJax's stylesheet as a regular stylesheet in the
project, and turns off the embedding. This is not the ideal way to do
it; if a MathJax version update or config change changes what should be
in that stylesheet, the change won't get applied properly. It is the
simplest way, though, and will solve several of the more pressing
load-time problems.

(Posts will only update to get rid of the duplicate stylesheets if
they're opened in the editor and re-saved.)