Change format of logoUrl from svg to png
The image pointed to by logoUrl is, as far as I can tell, used only in
emails. (Not that we're sending much in the way of emails yet, but we
will be.) Unfortunately, gmail does not support the svg format at all.
Presumably it's a security paranoia thing. Change logoUrl to be a PNG
instead, so that when it shows up in emails, it isn't broken.