First-pass email subscription from the front page
When you click the email subscribe button and then click Subscribe, it
does indeed subscribe you to curated. But this is a rough first pass;
the UI interaction is unnecessarily confusing, and the code is sort of
trying to be general to multiple different feeds, but actually only
supports Curated.