Don't clone comments while unflattening them
When comments are loaded into PostsCommentsThread, they come as a bunch
of comments with parent IDs, but they need to be a tree. The old code
(which we inherited from vulcan-forum) did this by cloning the comments,
then modifying them to have an extra chidlrenResults field. This messes
with the ability for components downstream to detect when things are
unchanged and don't need to be rerendered, leading to performance
problems. Change the way unflattening works, to a functional system
where instead we create an object with item: and children: fields,
leaving the comment itself unchanged.