Fix vote-count tooltip on posts
The vote-count tooltip on posts (but not comments) was trying to use
allVotes, which is not part of the fragment anymore, when it should've
been using voteCount. This was producing an error in the console, and
making it so that when you hovered over the score on a post, it didn't
tell you the vote count.