Sever inheritance relations coming from `replaceComponent`
If a component exists in example-forum and you're creating a component
with the same name in lesswrong to replace it, you can either use
registerComponent or replaceComponent. If you use replaceComponent, then
the addons from example-forum's call to registerComponent come along;
whereas if you use registerComponent, they don't.

This patch makes the lesswrong version of all components that override
example-forum components use registerComponent, so that deleting the
example-forum version will be a no-op.