If you aren't logged in, /account tells you to log in
The Edit Account (/account) endpoint is where you go if you click the
"Change your notification settings" link in an email. Previously, if you
weren't logged in, it would say "Sorry, you do not have the permission
to do this at this time." which is a bad thing to say to someone who's
trying to unsubscribe. Now it instead says "Log in to edit your
profile."