Restore portion of Bootstrap which handles Modal
The Modal component from react-bootstrap is used by Vulcan's
Components.ModalTrigger, which is in turn used for the "Please log in"
link below comments, and for "Register new RSS Feed" on user profile
pages. It was broken when we took out Bootstrap's CSS. This puts back
the specific subset of Bootstrap CSS rules which are used by Modal, so
that it works again.