Fix tab-hanging crash in Algolia search bar
SearchBox takes a `resetComponent` which is a React component. Under
some circumstances, React ends up doing a deep-compare between two
instances of this, where the instances being deep-compared contain
references to a `FiberNode` that owns them, which are deep and cyclic
and change while you're looking in such a way that the deep-comparison
never finishes.