Disable an eslint warning that was insane
By default, eslint warns if you use an apostrophe in JSX text (without
wrapping it in a string literal). The `eslint-plugin-react` thread about
it is some typography geeks arguing about the unicode standard, and
reaching the conclusion that apostrophes are actually U+2019. This is
clearly insane, so disable that warning.

https://github.com/yannickcr/eslint-plugin-react/issues/894