Add page titles to most endpoints
This adds titles to most routes. First, in Layout.jsx we add handling
such that putting title: on a call to addRoute works. Then, we add
titles to most of the routes.

This functionality overlaps with the stuff going on in
`getHeaderSubtitleData.js`, which is an ugly pseudo-switch-statement on
route names. We probably want to get rid of `getHeaderSubtitleData.js`
and replace it with similar extensions to addRoute.