Fix horizontal scroll on pages with the community map
The community map is set to be as wide as the browser window ("100vw").
Unfortunately, if the window has a Windows- or Linux-style scrollbar
(ie, one that isn't zero thickness), that's too wide by one
scrollbar-width, so it goes a little off the right edge, causing
horizontal scrolling.

This changes the map from `left:0` to `right:0` so that instead of going
a little off the right edge, it goes a little off the *left* edge. This
is better because the map's UI is all on the right (so missing 16px of
the left doesn't have any bad effects), and overflowing the left (rather
than the right) it doesn't create a horizontal scrollbar.