Type indicators in GroupLinks aren't buttons
The group type indicators (eg "EA", "LW") were implemented as buttons.
That meant they had the cursor style of a button, and a little click
animation. But clicking them didn't do anything. Make them not pretend
to be buttons.

(This has the side-effect of changing their font, which was previously
set by IconButton.)