Replace use of Components.ModalTrigger with new LoginPopupLink component
Components.ModalTrigger comes from vulcan:ui-bootstrap, which in turn
depends on Modal from react-bootstrap. This is a dependency we want to
be rid of, and also it was a bit hacked up (we fixed it after the
bootstrap removal by putting bootstrap's Modal-dialog styles in
_modal.scss).

Meanwhile @material-ui has its own Modal component, which does
approximately the same thing as react-bootstrap's (modulo a few
arbitrary style choices). Migrate over to @material-ui's Modal, filling
in the ModalTrigger portion with a small new component LoginPopupLink.