Remove react-boostrap `Alert` component from `PostsList`, `UsersPostsList`
React-bootstrap has a component `Alert`, which adds some styling--or it
would, if we had kept its CSS, but we didn't. So wrapping things in
Alert wasn't really doing anything except add a dependency on
react-bootstrap. So, get rid of that. Maybe we should put some styles
in, but on the other hand it's an error-handling path which is fairly
hard to trigger.