Re-remove Google Maps extra close button
Google Maps has a builtin close button, which is in the same spot as our
custom close button. We have some CSS to get rid of it. It came back,
because the upstream library changed it from a <div> to a <button>,
making our selector no longer match. Make it match again, and also add a
now-necessary !important.