Fix sequences navigation button coloring
There was a regression introduced by deleting seemingly-unused SCSS,
because `SequencesNavigationLink` was dynamically constructing the class
names `sequences-navigation-top-left` and `sequences-navigation-top-right`.
Additionally, the SCSS that was incorrectly removed contained a bug,
where the style that was supposed to apply to a disabled button (graying
it out) wasn't applied. Fix all that and move it to JSS.