Move all index declarations into view files
The goal here is to set up a convention where wherever a view is
declared, the indexes that view is likely to use appear in the same
spot. This obviously requires that indexes be declared in the same files
as the views, rather than separately.

This commit moves indexes around so that they're in view files, without
adding or deleting any.