RSS: Use a more-readable format for timestamps
The new format matches that of the tooltip that's displayed when you
hover over the date on the post page. I think the time zone will usually
be UTC for cloud RSS readers, though that's already the case.

Before: Published on Thu Nov 29 2018 09:32:54 GMT-0800 (PST)
After: Published on November 29, 2018 9:32 AM PST