Push tooltip withTimezone into a place that doesn't have to execute
The `withTimezone` HoC triggers some rerendering on page load (because
of switching from the server time zone to the client time zone), and
also date formatting with momentjs is a bit slow. Take the
date-formatting that went into `FormatDate`'s tooltip, and refactor it
so that it only has to happen (and only has to have a `withTimezone`)
if the tooltip is actually visible.