Fix a height inconsistency in SearchBar
When you clicked on the search button, the search area's height would
change by a couple pixels, making the login button next to it move
vertically. This layout brittleness was also causing a
more-than-a-few-px glitch in another change, which is in the next
commit.

Also, rename classes so 'root' is really the root.