Fix #1459: Search box doesn't autofocus
Algolia's SearchBox widget has a property `autoFocus`, which causes it
to take the keyboard focus on mount. However, we were mounting the
search box immediately, even when it's hidden with `display:none`, so
"on mount" was "on pageload" (a time when it can't be focused), rather
than on open.

Make SearchBox only be in the React tree when it's actually visible.
This causes it to focus correctly when it's opened.