Record user-agent and referrer with posts
The post/comment schemas we inherited from Vulcan/Example-Forum already
had fields for a referrer and a user agent, but they weren't being
filled in. The reason they weren't being filled in was because the
Vulcan mutator callbacks weren't providing callbacks with the necessary
information.

This is a pair of commits/PRs, one against Vulcan which adds that
information, and one against LessWrong which uses it to fill in referrer
and userAgent fields on posts and comments. This isn't used anywhere on
the site directly, but is something we want for analytics; most notably,
we want to know where comments/posts are posted from so we can better
distinguish between LessWrong users and GreaterWrong users.