Move callbacks to server-only
lib/modules/callbacks.js is full of callbacks that are actually only run
on the server (because they're mutation handlers). Some of them call
voting-related functions, which need to be server-only to be able to
import Algolia things. So, this is the first of a multi-step switchover:
first make callbacks server-only, then make some of the voting functions
server-only, then make them update Algolia.