Don't omit 'validate' option in various mutator calls
Vulcan's mutation functions have a named argument "validate", which was
never meant to be optional. However, due to the way it's used (as a
boolean with truthiness), it was effectively treated as an optional
parameter with a default value of false. No code within Vulcan relies on
this, but some code within LessWrong does. Fix that dependency, by
always providing the argument to `validate`. Paired with a Vulcan PR,
which will change the default value from falsy to truthy.