Convert custom_fields schemas from array to object style declaration
Vulcan's colletcion.addField has an awkward property, in that it can add
multiple fields at once, but only by using an array and an awkward
object format to handle the field-name/field-value distinction. So we
had a lot of schemas which looked like
  {
    fieldName: "foo",
    fieldSchema: {
      //...
    }
  },
When they should look like
  foo: {
    //...
  }
This adds a utility function addFieldsDict, and uses it to convert a
bunch of code from the former style to the latter style.

In order to make git merging work, this doesn't change the indent level
of the schema definition contents; it leaves those indented as-is so
they provide context for git to use to recognize what's what. They'll be
re-indented in the next commit.

This introduces one lint error, a duplicate field declaration warning.
That's a legitimate, pre-existing bug, which this made ESLint be able to
detect. It will be addressed in a separate commit.