Handle localStorage without library, different error handling
Not being able to access localStorage produces lots of noise in Sentry,
since many mobile devices don't have it or don't allow access to it, but
they're heterogenous in how they report this. So, strip away the
local-storage wrapper library (which wasn't doing anything useful), and
make localStorageHandlers catch everything and log it at only `warn`
level, so it doesn't go to Sentry.