Add a wrapper around react-router
This doesn't directly do anything; rather, it prepares us for a future
change which would otherwise be merge-conflicty. By taking the
merge-conflictiness up front.

As part of the Vulcan 1.13 upgrade (which we want mainly for the
attached Apollo 2.x upgrade), we're going to be switching from
react-router v3 to react-router v4. This is extremely merge-conflicty,
because react-router is used in a bajillion places, and the way it's
imported has changed.

Concentrate all of react-router's usages into one file
(reactRouterWrapper.js), so that we can make changes there, rather than
all over the codebase. That way we'll be able to have an Apollo2 branch
that we experiment with for a bit, without it accumulating merge
conflicts.