Move legacyData to separate collection, drop unused
This makes the Posts, Users, and Comments tables smaller, by (a)
dropping some fields inherited from Vulcan's example-forum which weren't
used, and (b) moving the legacyData field (which is large) to its own
collection. Features two new migrations, which can be run in either
order.

IMPORTANT: There is one actual use of legacyData in code, which is when
a legacy account logs in using their old-LW password. This has been
updated to use the legacyData collection so it should in theory work,
but it needs to be tested.