Add streaming support to meteor server-render
Fixes a bug in server-sink.js/appendContent where, if the thing you're
appending is a stream, it replaces instead of appending. Replace string
concatenations in server-register.js, which would cause a stream to turn into
the string "[Object object]", with calls to the combined-stream2 library
(preexisting dependency) to set up a stream that plays back one followed
by the other.