Activate lint for vulcan-accounts and make it pass
This involved fixing a minor bug in the `emailPattern` regex. It escaped
some '.'s, in order to try to enforce a URL-like structure (ie
foo@bar.com, must have a dot to the right of the @), but had the wrong
number of backslashes so it ended up just matching any character. Fix
that to get rid of the warning.