Speed up logged-out SSR
The way Users.owns was written, if the user is null (logged out), it
will dereference null, throw an exception, and then immediately catch
it. This is a mildly expensive operation, and Users.owns gets called a
lot of times. The practical upshot of which is that, for logged out
users, the server was spending 350ms of front-page server-side rendering
time just in Users.owns, throwing and catching internal exceptions.

Convert that into a proper null check, reducing the time spent in this
function to approximately zero.