Add /admin/migrations route
For admins only, /admin/migrations shows a list of all available
migrations, sorted by date written, whether they've been run and what
the result was. Rows are expandable to get a list of all runs of that
migration and when they were run.

There is a small new utility function, defineQuery in serverGraphqlUtil,
used for the backend. Later commits will refactor other things to use
that utility function.