Remove "searchable" option from schemas
Inherited from Vulcan, we had a notion of fields being "searchable".
What that actually means is that in the terms of a query going through
the view layer, you can provide a regular expression for that field. We
don't ever use this (though vulcan-admin does, via DataTable, but
DataTable is already pretty broken for unrelated reasons). If we ever
did use regex queries, it would bring the site down, because regular
expressions aren't (with the exception of prefix indexes) indexable.