Minimally convert lib/modules to TypeScript
The lib/modules folder didn't have a real distinction from lib, so it
had little reason to exist. Move everything that was in it to more
sensible places, and at the same time, rename them to Typescript (so
there's only one set of merge-conflicty updates of imports). Then does
the bare minimum to make type-checking pass.