benchmarkIncrementalStart: Slow down initial delay
Sometimes the server would not notice that a source file change had occurred
immediately, and would answer the benchmark's request with the old
version, giving a spurious almost-instant response. Lengthen the initial
delay, to prevent this from happening. Also remove the pre-test
is-server-up check, because that may have been slowing down the server's
noticing that a change had happened.