Remove Utils.getCollectionNameFromTypename
Utils.getCollectionNameFromTypename is a funciton which duplicated the
functionality of getCollectionName, but in a less-general way which
hardcoded answers for a specific subset of collections. It was used only
in SubscribeTo, and appears to have existed in order to support a
special case in the name mapping for a collection named "Categories"
which example-forum had but which we don't (and which needs a special
pluralization rule). Fixes one type error, and removes a weird trap
which would've triggered and caused a weird bug when we implemented
subscribing to tags.