Configure all settings in algoliaConfigureIndexes
Brings algoliaConfigureIndexes into sync with the configuration we have
on the production Algolia indexes, such that when I make a new index and
flip back and forth between its config page and production's config
page, they all match. This is without particular opinions about whether
those settings are *correct*, just that they're closer-to-correct than
the newly-created index defaults would be.