Fix a concurrency issue that could cause purged-user posts to remain in Algolia
The vote-cancel functions would save a copy of a document, do some
stuff, then update the document using their saved copy and some score
changes. But when purging a user, there were enough missing awaits slash
things happening concurrently that, in between when cancelVoteServer
saved a copy of the document and when it used it to update the index,
its copy would be stale, so that its index-update was effectively
reverting an index-purge that happened elsewhere.