When loading the logged-in user for resolver context, reuse it for the currentUser
When you make a request, we look up your user object based on the login
token. Save that user object to the dataloader, so that when the
currentUser resolver runs, it doesn't have to make a roundtrip to the
database. Speeds up SSR by one database roundtrip time (significant in
development, less significant in production).