All EA changes since last merge
Due to an error in the last merge, we needed to recreate it. Now we need to apply all the changes since then. I don't know how to do that with cherry picking, because of reasons involving branching. So I'm applying the diff since then in a patch.

See the changes in their original form here: https://github.com/centre-for-effective-altruism/EAForum/compare/6879002bc66325386cf9b933b2d91a56c2d9b337...ea-forum-devel