Move schema translation to initGraphQL
Move the bits of code that generate an actual graphql schema to a
server-specific file. This temporarily creates a nasty circular import,
because getSchema() actually *modifies* the query/resolver/mutation
lists (adding generated queries/resolvers/mutations to it) as a weird
side effect. Side effect to be cleaned up later.