Remove queryLimitName from RecentComments
This feature added the load-more quantity to the URL parameters when you
clicked Load More on the user profile. Unfortunately, it also caused the
comments list (and an unrelated posts list on the same page) to refresh.
Since this feature is unimportant relative to the impact of the broken
loading state, take it out.