Clean up voting API
Before: mutation vote takes a collection name, a voted-on document ID,
and a vote type, and returns a document with updated scores as a union
type. To cancel a vote, "cast" a vote of the same type as the user's
last vote, toggling it. To see if the user has voted on a document,
voteable collections have a currentUserVotes resolver, which returns an
empty array (not voted) or an array with a single vote in it (with vote
type, power, _id, etc).

After: The vote mutation and currentUserVotes resolver still exist, but
are deprecated. Instead there's a setVote[type] mutation for each
voteable collection, eg setVotePost, which takes a voted-on document ID
and a vote-type, and returns that document with scores updated, as the
appropriate type (with no need for a union). To cancel a vote, set the
vote type to null. To see if the user has voted on a document, voteable
collections have a currentUserVote resolver, which results in either
null (not voted) or a string (the type of vote which the user cast).