Mass-convert mongodb operations to async
Converts almost all calls to mongodb operations to async. In Meteor,
under the hood, they're still using fibers and blocking; they just have
the awaits and type annotations so that if they became truly async, it
would (in principle) keep working.

Major caveat: probably breaks some subtle aspects of account
creation/login, because of the users-collection callback hacks not
handling async stuff right.

Things not covered, which are still sync or a mix of sync and async:
bulkWrite; aggregate; a few get-post calls deep in Vulcan helpers which
bubble out into components and are reasonably stubbable but very hard to
convert; ban-check callbacks and legacy-login; startup initialization of
database settings.