Unbreak LWTooltip
LWTooltip had a "clickable" prop, defaulting to true. But LWTooltips
were never clickable because they had an event handler which made it so
that if you hovered over them, they'd disappear or flicker. Change the
prototype to reflect the reality, which is that LWTooltip is never used
for clickable things, and remove the horrible stopHover hack.