Change handling of canonicalSequenceId to fix some broken sequence navigation
Before: If both canonicalNextPostSlug/canonicalPrevPostSlug and
canonicalSequenceId are present, then for purposes of finding the
next/prev post, canonicalSequenceId is ignored. If the canonical
next/prev slug points to a post that doesn't exist or is inaccessible
(such as a draft), then is no next/previous link.

After: canonicalSequenceId takes precedence over canonicalNextPostSlug/canonicalPrevPostSlug,
which are only meant to be used for bridging the gaps at the start and
end of sequences which connect to each other. If canonicalSequenceId is
invalid or can't be used to find a next/prev post.