Fancy console logging for analytics events
Problem: Analytics events don't get much testing, because they're hard
to see, creating data-quality issues. But dumping them all to the
browser console would be really spammy and hard to read. Ideally there
would be a way to look over the analytics events going by which is
formatted in a way that's compact, and  makes it easy to parse what's
going on.

Solution: Fancy browser console-logging APIs. We show analytics events
in a single line, with an expander. The event-type string is hashed to
generate a color, so events are color-coded and easy to skim. Note that
this is only turned on if isDevelopment is true (ie, not on the
production site).