Log field changes to LWEvents
This is motivated by occasionally seeing things in weird states, and
wanting to know the history of user/moderator actions. Eg, was this post
moved back to drafts by the user, or by a moderator, and if a moderator
then which one? Has this user had their name/slug changed? Did this
wiki-tag used to have a different name, and who renamed it?

This solves the problem in general, by logging most changes (that go through
mutators) to fieldChange events in the LWEvents collection. Fields that
are marked as denormalized are excluded, as are content fields (which
are already versioned through the Revisions collection) as are a few
other miscellaneous things.

No user interface is provided for viewing this information, it's just
database logging.