Fix styling on the Concepts page
This got broken by a CkEditor upgrade, which changed the way it
expresses the `style` attribute on `table`, `td` etc to use a different
subset of the CSS `border-*` attributes. The new styling would generate
the same appearance, except that some of the attributes aren't in the
HTML validator whitelist, so they got strippedo ut on save. So add
`border`, `border-color`, and `border-style` to the HTML validator
whitelist so they go through. The appearance of the Concepts page will
be fixed on the next edit-and-resave after this merges.