Editor doesn't touch vulcan-forms or the document directly, only the wrapper
Makes it so that Editor has something more like a proper interface: all
interaction with vulcan-forms is handled by the wrapper, as is all
interaction with the document object.

As a prerequisite for part of this, add forwardRef support to
vulcan-lib/components. This allows components that have HoCs to hold a
ref and have their methods called by the parent. The wrapper uses this
to extract the serialized document from the editor when the form is
submitted.