Add validation of legacy IDs to comment hover preview and redirect
Errors related to this appear in Sentry, where the legacyIdComment view
is somehow getting passed a legacyId string which is the Unicode LTR
marker. (%e2%80%8e). Probably because of some weirdly malformed Markdown
somewhere deep in the old-LW archives. I speculate that this is one of
the sources of recurring 5xx errors in our monitoring. I haven't
actually reproduced it or found the malformed link, though.