Clean up and log in Sentry if there are core dumps
We had an issue where servers would leak memory until they ran out,
crash, and get restarted by ESBuild. Each time this happened, it would
leave a 3GB core dump inside of the Docker container; eventually, this
would cause the instance to run out of disk space, and no longer be able
to restart.

First: Record an error in Sentry if we notice a core file is present
when starting the server, since that indicates the server may have died
in a way where Sentry didn't get a chance to record the cause of death.
And second: remove the core files. In theory it may be possible to get
useful information out of them using lldb+llnode, but I gave that a
serious effort and couldn't get it working, so I think the core files
are actually just wasted disk space.