Show all tags and Create tag in Add Tag menu open in new tab rather than inline
The inline-in-the-menu UI made some sense when there weren't many tags
and we wanted to nudge people into populating them, but no longer makes
sense, and also it's a bit broken. In particular, clicking Show All Tags
would give you a UI that doesn't work at all on mobile, actually shows
only up to 500, and works pretty badly on desktop; and clicking Create
Tag would give you an inline editor that was eager to close if you
clicked anywhere.

Co-authored-by: jpaddison3 <jp@centreforeffectivealtruism.org>