Fix Recent Discussion feed failing to load when no posts
The Recent Discussion feed has a couple widgets (eg
RecentDiscussionSubscribeReminder) which are mixed in at numeric offsets
(ie slot #6), when the list is otherwise sorted by date. This caused
that section to fail to load and show an error message if there weren't
enough posts for that slot to exist.