Fix type error in MuiTableCell styling override
This was spreading `typography.fontFamily` into the default theme's
`overrides.MuiTableCell.body`, but `typography.fontFamily` is a string,
not an object, so this doesn't do anything sensible. I believe this had
no visual effect, but also MuiTableCell is only used in some fairly
obscure places, so taking it out is pretty low risk.