Fix bug that caused GreaterWrong posts to not appear in Latest Posts
Posts submitted through GreaterWrong, for some reason, wind up with
`isEvent` set to `null` rather than `false`. This causes them to be
excluded from the Latest Posts list, and from the list of posts on user
profile pages, because they filter out events with `isEvent:false` and
the `null` doesn't match.

Fix by checking for isEvent being null and changing it to false. I'm not
sure why GreaterWrong does this, probably there's more trouble to come
here. Our API probably should be enforcing non-nullness for graphql
fields of type Boolean *in general*, but it doesn't, and fixing that
would be a larger change.