Add a hide from frontpage button to posts
This creates a "Hide/Unhide post from frontpage" button
in the context menu of posts on the "Frontpage posts" and "All Posts"
(to allow users to unhide) which will, as you'd expect, hide a post.

This is a request from super users: sometimes they visit the forum quite
often, and find their frontpage has the same content, making it hard
to discover what's new. This allows users to filter their frontpage a
bit, making it easier for them to find what's new.

This button is only shown on the "Frontpage posts" and "All Posts"
sections, and only hides posts in the "Frontpage posts" section. This
hopefully makes it more clear that this won't ever e.g. let users hide
pinned posts, and lets user recover an accidentally hidden post by going
to all posts.

There's three major things I'd have included if I had the time, tl;drs
here, longer notes in the code:
1) Revisit the data model. The current data model is fine, but could
be a problem if users hide a lot of posts. A model similar to
`ReadStatuses` is likely preferable.
2) Fix race condition in mutation. Unlikely to be hit in practice, but
would be nice to always be reliable here.
3) Move filtering to server side. If users hide a lot of posts, the
frontpage becomes empty, or the pages become small.