Extract presentational components from DistanceUnitToggle
We want to reuse toggle buttons in the same style of the
DistanceUnitToggle. Therefore this commit extracts two components and
gives them an interface similar to the respective component in
material-ui.

In material ui v3 the ToggleButton(Group) is not yet
viable, but when this repository has been migrated to mui v5, then using
the stock ToggleButton(Group) should be strongly considered.

Also, the DistanceUnitToggle is not only used in the community page but
also on the events page and maybe even in more places. It should
probably move to a more shared location, but that can happen in a new
commit.