Always show filter for in-person/online events
This replaces the Select for in-person / online / all events with a
ToggleButtonGroup in the same style of the distance unit selector as a
select is usually not ideal for just three options. It hides the
available alternatives and requires the user to click in order to see
them.

Also, this no longer applies the class that hides the element on small
screens, as especially on mobile the user is more likely to be only
interested in in-person events, and always showing all events can add a
lot of noise to their experience.