Address corner cases in auto-refresh
Don't double-refresh if we get another build ID while a refresh is
already in progress. Try to ceconnect to the websocket after
disconnection, so that autorefresh will continue to work after killing
and restarting a server. When tabs are waking up, don't tell them to
refresh immediately if in the middle of a server restart.