Remove forum-gates in components.ts.
Removes unnecessary forum-gating from components.ts, and adds a
block-comment explaining how that file works and why forum-gating there
isn't necessary. This also fixes a spurioous warning on LW about the
`TargetedJobAdSection` component not being found (that component was
forum-gated in both components.ts and at the place where it's used, but
only the latter is necessary).