Fix spurious "Async before callbacks should not return undefined" console errors on client-navigate4
The handling of client-side page navigation event hooks (useOnNavigate)
was using Vulcan's callback mechanism, in a way that was intended for
callback-chains returning a series of modified versions of a document
(but there was nothing like that to modify). Some recently-added
defensive programming added a console error for `undefined` showing up
in those sorts of callback chains, which produced spurious console
errors on page-navigation.

Refactor to use a boring array for managing callbacks, fixing the
spurious console errors.